# swagger.yml
swagger: "2.0"
info:
  description: "API for managing trails in the Trail Management System."
  version: "1.0.0"
  title: "Trail Management API"
host: "127.0.0.1:5000"
basePath: "/"
schemes:
  - "http"
paths:
  /auth:
    post:
      summary: "Authenticate User"
      description: "Authenticates the user and returns a JWT token."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            required:
              - email
              - password
            properties:
              email:
                type: "string"
              password:
                type: "string"
      responses:
        200:
          description: "Login successful"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              token:
                type: "string"
              role:
                type: "string"
        400:
          description: "Email and password are required."
        401:
          description: "Authentication failed."

  /trails:
    get:
      summary: "Get All Trails"
      description: "Fetches all trails from the system."
      produces:
        - "application/json"
      responses:
        200:
          description: "A list of trails."
          schema:
            type: "array"
            items:
              type: "object"
    post:
      summary: "Create a New Trail"
      description: "Allows admin to create a new trail."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          name: "Authorization"
          required: true
          type: "string"
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            required:
              - Trail_name
              - Trail_Summary
              - Trail_Description
              - Difficulty
              - Location
              - Length
              - Elevation_gain
              - Route_type
              - OwnerID
            properties:
              Trail_name:
                type: "string"
              Trail_Summary:
                type: "string"
              Trail_Description:
                type: "string"
              Difficulty:
                type: "string"
              Location:
                type: "string"
              Length:
                type: "number"
              Elevation_gain:
                type: "number"
              Route_type:
                type: "string"
              OwnerID:
                type: "integer"
      responses:
        201:
          description: "Trail created successfully."
        400:
          description: "Missing required field."
        403:
          description: "Unauthorized access: Admin privileges required."

  /trails/{trail_id}:
    get:
      summary: "Get a Trail by ID"
      description: "Fetches details of a specific trail by ID."
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "trail_id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Trail details."
        404:
          description: "Trail not found."
    put:
      summary: "Update a Trail by ID"
      description: "Allows admin to update a trail by ID."
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "header"
          name: "Authorization"
          required: true
          type: "string"
        - in: "path"
          name: "trail_id"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              Trail_name:
                type: "string"
              Trail_Summary:
                type: "string"
              Trail_Description:
                type: "string"
              Difficulty:
                type: "string"
              Location:
                type: "string"
              Length:
                type: "number"
              Elevation_gain:
                type: "number"
              Route_type:
                type: "string"
      responses:
        200:
          description: "Trail updated successfully."
        400:
          description: "Missing required field."
        403:
          description: "Unauthorized access: Admin privileges required."
        404:
          description: "Trail not found."
    delete:
      summary: "Delete a Trail by ID"
      description: "Allows admin to delete a trail by ID."
      produces:
        - "application/json"
      parameters:
        - in: "header"
          name: "Authorization"
          required: true
          type: "string"
        - in: "path"
          name: "trail_id"
          required: true
          type: "integer"
      responses:
        200:
          description: "Trail deleted successfully."
        403:
          description: "Unauthorized access: Admin privileges required."
        404:
          description: "Trail not found."